git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1140 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/runtime/RexxBuiltinFunctions.h
@@ -118,6 +118,6 @@ RexxString *PackHex(const char *String, size_t StringLength);
 /*********************************************************************/
 
                                        /* convert the number                */
-#define IntToHexDigit(n) "0123456789ABCDEF"[n];
+#define IntToHexDigit(n) "0123456789ABCDEF"[(int)n];
 
 #endif
